# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "用于知夕番茄钟程序的后台api文档",
        "title": "知夕后台Api接口",
        "contact": {
            "name": "Doremy",
            "url": "https://dimdoremy.github.io/"
        },
        "license": {},
        "version": "1.0-golang"
    },
    "host": "127.0.0.1:8848",
    "basePath": "/api",
    "paths": {
...
        "/deletehandbook": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "book-routers"
                ],
                "summary": "删除图鉴数据库的指定条目",
                "parameters": [
                    {
                        "description": "删除用户信息",
                        "name": "handBook",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.HandBook"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "$ref": "#/definitions/model.MessageBind"
                        }
                    },
                    "400": {
                        "description": "400 bad request",
                        "schema": {
                            "$ref": "#/definitions/model.MessageBind"
                        }
                    }
                }
            }
        },
...
        "/deleteuserpiece": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "piece-routers"
                ],
                "summary": "删除背包数据库指定条目",
                "parameters": [
                    {
                        "description": "单个用户信息",
                        "name": "handBook",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UserPiece"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "$ref": "#/definitions/model.MessageBind"
                        }
                    },
                    "400": {
                        "description": "删除失败",
                        "schema": {
                            "$ref": "#/definitions/model.MessageBind"
                        }
                    }
                }
            }
        },
        "/handbookbyopenid": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "book-routers"
                ],
                "summary": "获取单个用户图鉴列表信息",
                "parameters": [
                    {
                        "description": "单个用户信息",
                        "name": "handBook",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.HandBook"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "单个用户图鉴信息的json数据",
                        "schema": {
                            "$ref": "#/definitions/model.HandBook"
                        }
                    },
                    "400": {
                        "description": "400 bad request",
                        "schema": {
                            "$ref": "#/definitions/model.MessageBind"
                        }
                    }
                }
            }
        },
...
        "/inserthandbook": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "book-routers"
                ],
                "summary": "注册新用户新建图鉴信息",
                "parameters": [
                    {
                        "description": "注册用户信息",
                        "name": "handBook",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.HandBook"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "注册成功",
                        "schema": {
                            "$ref": "#/definitions/model.MessageBind"
                        }
                    },
                    "400": {
                        "description": "注册失败",
                        "schema": {
                            "$ref": "#/definitions/model.MessageBind"
                        }
                    }
                }
            }
        },
...
        "/insertuserpiece": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "piece-routers"
                ],
                "summary": "注册单个用户，新建背包",
                "parameters": [
                    {
                        "description": "单个用户背包信息",
                        "name": "handBook",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UserPiece"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "注册成功",
                        "schema": {
                            "$ref": "#/definitions/model.MessageBind"
                        }
                    },
                    "400": {
                        "description": "注册失败",
                        "schema": {
                            "$ref": "#/definitions/model.MessageBind"
                        }
                    }
                }
            }
        },
...
        "/piecebyopenid": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "piece-routers"
                ],
                "summary": "获取单个用户背包信息",
                "parameters": [
                    {
                        "description": "单个用户信息",
                        "name": "handBook",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UserPiece"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "单个用户信息的json数据",
                        "schema": {
                            "$ref": "#/definitions/model.UserPiece"
                        }
                    }
                }
            }
        },
...
        "/updatehandbook": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "book-routers"
                ],
                "summary": "更新单个用户图鉴信息",
                "parameters": [
                    {
                        "description": "更新的用户信息",
                        "name": "handBook",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.HandBook"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "$ref": "#/definitions/model.MessageBind"
                        }
                    },
                    "400": {
                        "description": "400 bad request",
                        "schema": {
                            "$ref": "#/definitions/model.MessageBind"
                        }
                    }
                }
            }
        },
...
        "/updateuserpiece": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "piece-routers"
                ],
                "summary": "更新单个用户背包信息",
                "parameters": [
                    {
                        "description": "更新的用户背包信息",
                        "name": "handBook",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UserPiece"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "$ref": "#/definitions/model.MessageBind"
                        }
                    },
                    "400": {
                        "description": "更新失败",
                        "schema": {
                            "$ref": "#/definitions/model.MessageBind"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.HandBook": {
            "type": "object",
            "properties": {
                "monster1": {
                    "type": "integer"
                },
                "monster2": {
                    "type": "integer"
                },
                "monster3": {
                    "type": "integer"
                },
                "monster4": {
                    "type": "integer"
                },
                "monster5": {
                    "type": "integer"
                },
                "monster6": {
                    "type": "integer"
                },
                "monster7": {
                    "type": "integer"
                },
                "openid": {
                    "type": "string"
                }
            }
        },
        "model.MessageBind": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "model.UserPiece": {
            "type": "object",
            "properties": {
                "donghuangzhong": {
                    "type": "integer"
                },
                "fuxiqin": {
                    "type": "integer"
                },
                "haotianta": {
                    "type": "integer"
                },
                "kongtongyin": {
                    "type": "integer"
                },
                "kunlunjing": {
                    "type": "integer"
                },
                "lianyaohu": {
                    "type": "integer"
                },
                "nvwashi": {
                    "type": "integer"
                },
                "openid": {
                    "type": "string"
                },
                "pangufu": {
                    "type": "integer"
                },
                "shennongding": {
                    "type": "integer"
                },
                "xuanyuanjian": {
                    "type": "integer"
                }
            }
        },
...
    }
}